Began refactoring MusicPlayer
