+++ src/main/java/MusicPlayerManager.java
@@ -1,0 +1,7 @@
+public class MusicPlayerManager {
+    
+    private MusicPlayer musicPlayer = new MusicPlayer();
+    
+    
+    
+}
